pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdbabrenum"
version = "0.1.0"
description = "A tool for renumbering antibody structures in PDB files with multiple numbering schemes"
authors = [
    {name = "yumizsui", email = "yumizsui@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["antibody", "pdb", "renumbering", "imgt", "chothia", "kabat", "bioinformatics"]

dependencies = [
    "biopython>=1.78",
    "abnumber>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
pdbabrenum = "pdbabrenum.cli:main"

[project.urls]
Homepage = "https://github.com/yumizsui/PDBAbRenum"
Repository = "https://github.com/yumizsui/PDBAbRenum"
"Bug Tracker" = "https://github.com/yumizsui/PDBAbRenum/issues"

[tool.setuptools]
packages = ["pdbabrenum"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
